+++ samples/Zeta.Sample.FastEndpoints.Api/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using FastEndpoints;
+global using Zeta;
+global using Zeta.FastEndpoints;